requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
cryptography==42.0.5
gunicorn==21.2.0
python-dotenv==1.0.1
